feat(grid3d): constant color for coloring
